Update patch set 2
Patch Set 2:

But only for modules where the makefile actually does specify incompatible flags (I believe this is a very small number of them). Adding it to the build system adds processing time to the makefiles (not sure how much, in practice these lists are small so probably not much). I'm more concerned with the fact that someone adding an imncompatible cflag to a module that uses clang expecting it to work will now have it silently ignored...

Patch-set: 2
Label: Code-Review=0