Update patch set 1
Patch Set 1:

I do agreed that it does not make sense to don't zipalign some entries and others not (despite DexGuard supports that). However, the main issue here is that despite re-zipalign an apk should be a noop, it is not the case while using DexGuard's tampering mechanism. 

I don't have the details of how they are implementing the tampering mechanism, but both, their documentation and my tests, show that if you re-zipalign an apk that has Dexguard tampering mechanism & zipaligning, it will fail.

Clearly, Dexguard relies on some kind of signature for their tampering mechanism, I assume similar apps providing tampering mechanisms will do the same; therefore a way to skip the zipaliging for PRESIGNED packages should be provided in order to avoid any post modification of the file during the build process.

Patch-set: 1