Update patch set 1
Patch Set 1:

davidxl: seems like we're missing a linker option here. (or i'm misunderstanding something.) we *do* want the linker to warn us if the .so we're creating is missing dependencies, but we *don't* want it to transitively check the libraries we're linking against.

aiui, the default is transitive checking and this option means no checking, and there's no way to get what we want?

(the problem is that the system libc and the NDK libc don't contain the same set of symbols. there's more in the system libc, and a system library might refer to these, but your NDK-built library can't (but it can refer to a system library that transitively pulls in the system libc symbols).)

Patch-set: 1