Update patch set 1
Patch Set 1:

note that --allow-shlib-undefined is the default behavior for GOLD linker, so this has always being the case for arm32. Gold linker does not attempt to track down needed libraries by design -- so at least we won't get this behavior (warning) from Gold.

For gnu-ld, for which --allow-shlib-undefined is needed, it might be useful to have warning as you suggested, but the actual implementation may be tricky and probably not worth the effort given that gold will be used in the future.

Patch-set: 1