Fix {java,native}-{host,target}{-tests} make targets for 2nd arch.
The expectation when the user runs e.g. 'make native-target' is that
all of the native code for the target is built. What actually happens
is that the primary architecture's code is built, but the secondary
architecture's code is not built.

This happens because these extra targets hijack the my_checked_module
variable which (as the name implies) is primarily used for the module
checking feature. For module checking, the second architecture is
intentionally disabled.

This fix copies the contents of my_checked_module to another variable
before the module checking workaround is applied, so the extra make
targets do the right thing.

This feature is useful for distcc so a high number of make jobs can
be passed to make for the native code, but a smaller number for the
java code in a second pass.