Update patch set 4
Patch Set 4: Verified+1

There's still one case where this doesn't (and probably can't) work: a static library built with coverage with gcc cannot be linked into an executable built with coverage with clang (or vice versa) because either libprofile_rt OR libgcov can be used, but not both. I don't think there's any way to solve this problem, so it will just have to be avoided.

Patch-set: 4
Label: Verified=+1