Update patch set 1
Patch Set 1:

Elliott, Ian, Brian,

The basic question here is - would we like 32bit ARMv8 builds to use instructions that are not available for ARMv7, but make things faster? 

My view here is that we should do this - we are building the system image, not an application that needs to be working across a variety of devices** (the same way we apply errata fixes only where needed :-) ).

For ART we can do the architecture variant detection at runtime, and if we run on an ARMv8 CPU then we can generate AArch32 specific instructions (not available in ARMv7).

For other places, like bionic, we need a build time flag that tells us that we are building for a ARMv8 32bit target (2ND_ARCH of a ARM64 build).

A good AArch32 optimisations example is libm's floor() - currently:
https://android.googlesource.com/platform/bionic/+/master/libm/arm/s_floor.S

all this can be as simple as:

mov s0, r0       // ARM ABI build is soft-float
vrintm s0, s0
mov r0, s0       // ARM ABI build is soft-float


** If we enable architecture variant specific instructions in libraries that endup in the NDK we should be really hide them away from the unexperienced user (if we expose them at all)! We do not want to end up in a situation where somebody builds a 32bit library that uses AArch32 only instructions (and thus not run on an ARMv7 device).

I hope this helps more than my initial comment,
Serban

Patch-set: 1