Update patch set 1
Patch Set 1:

> > (this is only for the art folks. we can't use aarch32 on the
 > native side.)
 > 
 > What do you mean like that? Seems like openssl/boringssl would want
 > to use this for building their native code? I don't recall how they
 > are taking advantage of the new instructions on 32-bit today,
 > possible through runtime CPU detection.

OpenSSL sets a handler for SIGILL and tries the ARMv8 instructions to see if they work. If so, it'll use them.

BoringSSL uses getauxval if it can to avoid that, but will fallback to SIGILL probing if needed. BoringSSL is already used in Chrome on Android and we've not had any reports of problems. (Although I'm not sure whether Chrome on ARMv8 systems is running 32 or 64-bit native code.)

Patch-set: 1
Reviewer: Gerrit User 1019036 <1019036@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0