Update patch set 5
Patch Set 5: Code-Review+1 Verified+1

-Wno-error=\#warnings is a clang thing. -Wno-error=cpp is a gcc thing.

Confirmed that this builds and produces the desired products.

Thanks for the rewrite, yes it is simpler if we accept android_filesystem_config.h _must_ be the basename of the device-supplied file.

(I expect that future-selves will wish we handled multiple_files, one file is good enough for today).

Patch-set: 5
Label: Verified=+1