Update patch set 1
Patch Set 1:

w/o zipalign as a related change and first example; is this change standalone the first in the series to allow fs_config to be allowed to use shared libraries. zipalign can continue using the static libraries (until the SDK problem is resolved as a _separate_ cleanup).

Patch-set: 1