Update patch set 2
Patch Set 2:

Bernhard, if that is the case, that would mean that a53 and a57 ought to have their own 2nd arch (not just an existing 32-bit chip) or they would be missing out as well, or would they be better off using Denver as CPU Variant?  Also, should -mfpu=crypto-neon-fp-armv8 be used instead to reflect the crypto extensions?

Patch-set: 2