Update patch set 2
Patch Set 2:

yes, a53 and friends should probably have their own 2nd arch variant (and the right -mcpu= setting attached to it), there are a couple of extensions to the 32bit instruction set in armv8.

I'm not sure if crypto is mandatory in v8; if it is, it should probably be enabled there.

I don't think gcc or clang currently generate crypto instructions by themselves, so AFAIK the only difference for now is that they would be legal in assembly code -- but of course compilers keep getting smarter (and actually I may already be outdated).

Patch-set: 2