RFC: ARM: Add support for more ARM CPU_VARIANTS.
This patch adds support for more ARM CPU_VARIANTS including 32bit and
64bit ARMv8. However there are still some issues related to building
32bit ARMv8 platforms - see the comments in armv8-a.mk.

Please note that this patch is a request for comments! This might not
be the preferred direction to take, but it aims to solve the current
issues in a more generic manner. Also note that there are still issues
with regard to 32bit AArch32 platform builds (ARMv8-A).

Change-Id: I474bf14b0313ca890a3342ea119119badcd2028b
Signed-off-by: Serban Constantinescu <serban.constantinescu@arm.com>