Update patch set 1
Patch Set 1:

To clarify, this complication is necessary due to the architectural limitation of ASan:
Libraries built with ASan can only be used (at runtime) with executables built with ASan. ASan executables, on the other hand, can load both ASAn and non-ASan libraries.
We also assume that we can not build all executables in the OS with ASan at once. This means that we need two sets of shared libraries, etc.

Patch-set: 1