Update patch set 1
Patch Set 1:

> > (2 comments)
 > >
 > > I am afraid this will bloat the system.img size too much. On what
 > > builds/devices do you want to enable this build?
 > 
 > Could we put all the instrumented .so on /data?
sounds like a good idea.
Can we build a overlay folder structure in /data that can be applied to a normal build?

 > I guess we will start with clusterfuzz devices, i.e. N5/6/9.
 > 
 > > Also did you ever try a build? I am not sure how well it
 > interacts
 > > with custom rules in Android.mk, though generated source files in
 > > $(local-generated-sources-dir) should be fine now.
 > 
 > I marked a bunch of targets for sanitization and ran build (32-bit
 > only though, for now), it succeeded but it looks like something
 > went slightly wrong as com.google.bluetooth is segfaulting at
 > startup, but everything else works fine. I'll investigate.
 > 
 > Could you point me to something you think could fail? I can't just
 > enable asan for everything without a significant cleanup.

I don't have a specific case. 
But in general if the intermediate dir of a module (by call to local-intermediates-dir or intermediates-dir-for) is directly referred to in an Android.mk,
or if an Android.mk refers to variables set up by the core build system after "include $(BUILD_SHARED_LIBRARIES)" to do postprocessing,
we'll have hard time auto-generating modules.

Patch-set: 1