Stop building all modules with tag tests by default
Stop building all modules with a tests module tag.
Forcing local builds to always build all tests are too costly.

On master branch this reduces what is built by more than
300 apps and 50 java libraries. Time for doing this on a
12 core machine running with -j13 is about 10 minutes.

It is important to catch build breakage among the test
projects but the cost of forcing all developers to do
this is rather high. It is more reasonable to have a
central build job set up that can do back-to-back builds
asynchronously to catch build issues, than have a lot
of people waiting for builds to complete.

The default make target dont build all packages anyway
it only build what is included via PRODUCT_PACKAGES so
it is a bit odd that test targets get included automatically.

Change-Id: I0cf18bc5ba44e26b01e785b3c3fad2550fe0333c