WIP: Split overlapping move commands
We may break a move that has overlapping blocks into multiple
non-overlapping ones, if a) it doesn't use any stash and b) both of its
src and tgt are continuous ranges. Non-overlapping moves don't need to
be stashed during an update, which improves the efficiency and reduces
stash space requirement.

For example, "move 1-10 to 4-13" now becomes "move 7-10 to 10-13; move
4-7 to 7-10; move 1-4 to 4-7".

Bug: 21202554
Change-Id: I3fb3d32f414d9aee01f67f8df7bb41a88e04adbb