LOCAL_DEX_PREOPT value is not accounted for second arch
prebuilt_internal.mk is called twice in case if the build module
supports two archs. For the second call, we need to pass the same
value of LOCAL_DEX_PREOPT, which is setup in Android.mk, to make
sure it is accounted properly. Otherwise, even if we set it to
false in Android.mk, dexpreopt is still invoked if that build
component supports libs for both archs (LOCAL_MULTILIB := both)