RFC: Implement BUILD_[HOST_]STATIC_AND_SHARED_LIBRARY
Build both a static and a shared library from the same module
definition.

This adds another layer of variants to some LOCAL_ variables:

  LOCAL_SRC_FILES_(static|shared)[_(32|64|arch)]
  LOCAL_STATIC_LIBRARIES_(static|shared)[_(32|64|arch)]
  LOCAL_SHARED_LIBRARIES_(static|shared)[_(32|64|arch)]
  LOCAL_WHOLE_STATIC_LIBRARIES_(static|shared)[_(32|64|arch)]
  LOCAL_CFLAGS_(static|shared)[_(32|64|arch)]

If possible, this will only build the sources once, and use the static
library as a whole static library in the shared library. If the cflags
are different, or the static library includes srcs/libs that the shared
library doesn't, they're build separately.

Change-Id: Idd6097ef9e873a7f9af7c2ff2eee775d5b4cf8fb