Update patch set 1
Patch Set 1:

> This looks like a reasonable workaround for llvm regression.
 > Please report it to llvm.org too and we can revert this change when
 > it is fixed upstream.

From the patch and the commit message, this seems the behavior intended by the change.  Whether it is reasonable, is another question.  We can discuss that tomorrow and file an upstream  feature request.

The problem is that if both -mcpu and -mfpu are used, the mfpu option should cover everything supported by the cpu.  For e.g., Krait supports neon-vfp4 and when mfp is set to neon, the behavior-change kicks in and the backend will not generate code specialized for Krait.

Just realized this also affects cortex-a{7,8,15} devices, for which we pass the -mcpu flag to the compiler.  The general solution would be to skip setting '-mfpu=neon' if we set '-mcpu'.

I think this is just a performance issue for non-Krait devices, with the backend not generating specialized code.  Nevertheless, we should update the build rules to skip -mfpu appropriately.  I'll file a bug in case we don't get to do it as a part of the rebase.

Patch-set: 1
Reviewer: Gerrit User 1058860 <1058860@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0