Update patch set 1
Patch Set 1:

I don't know if a better way exists, but I often need to find things inside Android.mk files, for instance, module names or references to other modules. I am not sure if this is the best way, but it works for me and I have been carrying it around for awhile. Uploading to see if others find it useful enough to pull in.

Patch-set: 1
Reviewer: Gerrit User 1013433 <1013433@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0