init.te: Add rw access to serial devices
This fixes the issue with the emulator "-shell" option.
Init tries to open the console which is passed through
the kernel androidboot.console property, but fails to open it.
This rule gives rw permissions to init to work with serial devices.

ARM64 Ranchu ttyAMA[0-9]* (pl011) has been labeled as serial_device
in file_contexts.

Change-Id: I9ef94576799bb724fc22f6be54f12de10ed56768