Update patch set 1
Patch Set 1:

Yes, this is because the jar tool is idiotic. "-C" options are global, they do not apply only for the following argument. For each entry in the jar file, it will compare the absolute path of the file
to *every* prefix specified by "-C" and arbitrarily choose the first that the absolute path starts with. So here, "external/smali/smali" applies to build.gradle, runantlr.sh and runjflex.sh.

In other words, it's best not to have "overlapping" "-C" options that are prefixes of each other.

This also implies that i'll just have to make copies of all resources i want. sigh.

Patch-set: 1