Update patch set 1
Patch Set 1:

> Does the backend mmap the file or do you crack it open with file-io
 > and store that somewhere? Wouldn't each process with libselinux
 > have its own data set, and does that dataset size increase with
 > .bin format?

For file_contexts.bin, we mmap the file read-only and directly use the precompiled regex from the mapping.  For reasons not entirely clear to me, we do copy the context string (ctx_raw) rather than directly using what is in the mmap'd file, but otherwise we seem to be directly reusing the mmap'd file contents.  So you get greater reuse of memory with file_contexts.bin than with file_contexts.

Patch-set: 1