Update patch set 1
Patch Set 1:

> I guess it depends then on whether the improvement in file contexts
 > load time is sufficient to justify switching to file_contexts.bin,
 > at the cost of more space.  init bootchart with and without
 > file_contexts.bin might be interesting, but there are also other
 > users of file_contexts that may benefit.

Actually, I guess init bootchart starts too late for this, plus loading file_contexts is likely in the noise there.  So we'd just need to put a couple of gettime() or similar calls around the call to selinux_android_file_context_handle() to measure the time to load file_contexts, and do it with and without file_contexts.bin present.

Patch-set: 1