Update patch set 1
Patch Set 1:

Actually, I think it makes sense to make the switch to file_contexts.bin even if it doesn't make a large difference in load time on Android, given that it also should save memory and that it will mean that we'll be exercising the same code path on Android as Linux (since Linux now always has a file_contexts.bin).  Only cost is additional storage for the file.

On Linux, one reason to keep the text file_contexts file around is to deal with the scenario where libpcre gets updated (independently of the policy).  libpcre doesn't provide any compatibility guarantees for the precompiled regexes; you are supposed to regenerate them when updating to a new libpcre version.  So we save the pcre_version() in the file_contexts.bin header when generating it and check it against the current pcre_version() when loading it, and fall back to the text file_contexts in that case.  As long as we can guarantee that /file_contexts.bin is updated whenever /system/lib/libpcre.so is updated in Android, then this isn't an issue there.

Patch-set: 1