Update patch set 1
Patch Set 1:

> How much ready-to-go componentry are their for the bin file_context
 > comparisions?

Well, none, but you would basically just take the label_file load_mmap code, load up the AOSP and device file_contexts.bin files, and then check that all of the AOSP entries are contained in the device one.  Shouldn't be hard to write but would have to be done.

Patch-set: 1