Update patch set 1
Patch Set 1:

The rules for combining sanitizers are actually more complex than I thought. ubsan can only be combined with asan, and only asan, msan, and tsan are mutually exclusive. This logic would be rather complicated to implement in make, and I'm not sure we really gain much from it. I'm going to just have the environment clobber the module rather than try merging the two. We can fix this later if we find a compelling reason to merge them (hopefully by then we'll be in soong).

Patch-set: 1
Label: Code-Review=0