Split large files for BBOTA v3.
For BBOTA v3, we need to stash source blocks to support resumable
feature. However, with the growth of file size and the shrink of the
cache size, source blocks that represent a file are too large to be
stashed as a whole. If a file spans too many blocks, we split it
into smaller pieces by generating multiple commands.

One potential downside is that after splitting, files like .jar,
.apk and .zip can no longer use imgdiff. We may lose the potential
benefit of using imgdiff for patch size reduction.

Bug: 22430577
Change-Id: I5684fab0ac41c8d94fe18f52a34cef77e06f741c