build: support compiling .mm on host
libchrome uses .mm (Objective-C++) files to bridge C++ code with
OS X Frameworks. This adds support for compiling .mm to .o using
the same approach as is used for .m to .o, except using the C++
compiler.

The C++ definitions are abstracted in the same way that the C
definitions already are to achieve this.

Bug: 24168923
Change-Id: Ia65357e2e2584dfffcb6796e214fe6b27635c3a6