Update patch set 1
Patch Set 1:

I started down the path of writing the tool that I needed to work on Mac in Python, but then we decided to adopt Binder going forward on Android, and the tool became less important.

I believe Scott wants these changes because for the time that we support Mac build, he would like his unittests to run on Mac, where he does his development.  This use case is sort of orthogonal to the tool I was working on.

Patch-set: 1
Reviewer: Gerrit User 1057759 <1057759@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0