[WIP/RFC] Use .KATI_RESTAT to prevent unnecessary rebuilds
Currently, when framework.jar is rebuilt, all apps are rebuilt
as they depend on framework.jar. With this patch, they are
rebuilt only when framework.jar's interface was changed.

With this patch, "m" after touching a java file in framework.jar
took 1m29s, rebuilding 215 targets. Without this, it took 4m18s
and built 1869 targets.

Change-Id: I39982c0cf50d99bfdfdde63c8371d8685b16450f