Add LOCAL_JNI_SHARED_LIBRARIES to required modules.
Previously the dependency of the apk on the jni .so files is set up only
in build/core/install_jni_libs_internal.mk:
$(LOCAL_INSTALLED_MODULE) : | $(addprefix $(my_shared_library_path)/,$(my_jni_filenames))
However, with the order-only dependency if the jni so files is updated,
make won't rebuild the system.img, which has dependency only on the apk
file.
With this change, system.img's direct depednency on the jni so files is set up.

Bug: 24865400
Change-Id: I9d5bee82b8a712a2c24dabaa0cd4c50174ea219f