Update patch set 6
Patch Set 6:

(2 comments)

I went and added -X to most of the zip use cases -- I didn't find anywhere that would actually be hurt from the less precise timestamps or lack of UID/GID of the builder.

If we ran remove-timestamps-from-package, more of them could be repeatable too, but that requires a temporary directory with the current zipalign, which isn't how most of these are set up right now.

The prebuilt path may still be a bit problematic -- the original SignApk would have set the timestamps to the first cert's begin date plus a month, but repackaging the libraries would still change the zip modification date for those libraries (though not the precise timestamp/uid/gid due to -X). But if we ran remove-timestamps-from-package we'd change the dates on every file, including the ones that SignApk touched.

Patch-set: 6