Always sign APKs and OTA update ZIPs using SHA-256.
Prior to this change, APKs and OTA update ZIP files were signed using
SHA-1 or SHA-256, depending on the signature/digest algorithm of the
X.509 certificate.

The signing/digest algorithm of the X.509 certificate has no relevance
to the security of APK or OTA update ZIP. Thus, this change switches
to signing using SHA-256 unconditionally.

Bug: 25643280
Change-Id: I47e84dfc255af5f0f6d797108ac3daf8f3720551