Update patch set 1
Patch Set 1: Code-Review-2

Order-only dependencies are pretty evil in my experience, so I would prefer that this not be the solution here. FWIW, arm_neon.h might look like it is pretty stable, but it in fact is not. They regularly change this in substantial ways, hence the build rule in the first place.

Perhaps the clang-tblgen rules are incorrect.

Patch-set: 1
Label: Code-Review=-2