Do not run zipalign when building APKs.
This changes the build system to no longer run zipalign except for
prebuilts which do not need to re-signed.

Running zipalign is no longer needed because signapk takes care of
alignment. For prebuilts which do not need to be re-signed signapk is
not invoked and thus zipalign is still needed because the prebuilts
may not have been aligned properly. This exemption will go away once
prebuilts are properly aligned.

Bug: 25794543
Change-Id: I31f10af80b9baa3c9481cb36f41e5e5046e852f9