Update patch set 1
Patch Set 1:

Yeah, it makes sense to land it for each architecture. Lots of obscure architecture and perhaps even device-specific issues.

I am pretty sure there is SSP for MIPS but it's fairly new. Not sure if it's actually present in the GCC version used by Android (might be). It would be worth testing it there, but it might just bail out from the switch not being supported yet.

Patch-set: 1