Update patch set 1
Patch Set 1:

the internal bug is basically "qualcomm's camera code is broken, so -fstack-protector-strong prevents you from using the camera". one limitation of the fact that i do my testing in AOSP is that i just get prebuilt blobs for stuff like that, so it's only in the internal tree where you can rebuild the vendor code with the new options that it becomes apparent.

Patch-set: 1