Detect COMMON_GLOBAL_C(PP)FLAGS changes
Reorder the definitions of the core cflags so that we can warn when
they are changed by the device configuration files. Once we clean up the
current devices, this can be changed into a warning, and eventually
dropped entirely.

There is no reason for devices to be configuring the global cflags -- if
something needs to change for a particular device, that flag should be
defined in the BoardConfig.mk, then used in the individual makefiles
where it is used. The common global flags also affected host tools,
which should not change between device configurations.

Bug: 26165350
Change-Id: Icece4505ce3d2a9168087d802953378534018d47