Update patch set 1
Patch Set 1:

(Same as on ART CL): Not sure what's best here. Could argue that if we can't get the dependencies that are newer (and so induce the update), we should list all dependencies then (to see what they are).

Patch-set: 1
Reviewer: Gerrit User 1041833 <1041833@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0