Update patch set 1
Patch Set 1:

Yeah, this won't help if anyone is using $(shell java*), since that environment isn't modified, but it'll help. I suppose we could move this up so that it's executed in the make instance instead of kati -- that's a much smaller environment, and would expose these modifications to kati itself, including $(shell). Thoughts Colin?

Patch-set: 1