Update patch set 1
Patch Set 1:

They are essentially shared libraries that we link against on the device. They have always been installed to /system/lib and rewriting the security rules to make this function from somewhere else is beyond my expertise. There are no other options for LOCAL_MODULE_CLASS that imply a shared library-like object, and I am hesitant to add a new class for something that there are very few of.

Patch-set: 1