Update patch set 1
Patch Set 1:

(1 comment)

I'd be hesitant too, but when we move these to soong, we'll be delivering them back to this build system using BUILD_PREBUILT, and then you won't want it assuming that it's actually a shared library.

I'm not actually proposing that we change anything the actual system image.

To add the class:
https://android-review.googlesource.com/202715
https://android-review.googlesource.com/202850

Patch-set: 1