Update patch set 1
Patch Set 1:

> It looks like we should be passing -fdebug-prefix-map= for assembly
 > files too, but maybe not for modules that ask for no default
 > compiler flags, or have custom rules. Did you see it in a
 > particular file, or for every assembly file?

it seems to at least be all the .S files in bionic.

(my usual debuggerd test is to run "crasher" from system/core/debuggerd. if you say "crasher abort" it will have a stack that includes assembler.)

 > I'll cherry-pick and
 > check this later.
 > 
 > And unfortunately, (at least in GCC) this argument gets added to
 > the producer in the dwarf symbols, so you can't just grep for the
 > path.
 > 
 > I should file a clang feature request to remove it from there, then
 > we could do something smarter with generated code in
 > absolute/nonstandard OUT_DIRs too. (though that may cause more pain
 > for debuggers)

okay, i'll submit this now (because it will fix the mess for most people), and abandon the script change on the assumption that you'll find a better solution.

Patch-set: 1