Fix C++ on Darwin
We've been including the system libc++ headers even if we're building against
our version of libc++. Stop doing that, and only add the headers to our path
if we're using the system libraries.

If nothing is specified, on recent OSX versions, libc++ is the default c++
library instead of libstdc++. We've been explicitly including the libc++
headers on all versions, but that breaks old versions. Force us over to libc++,
since the system libstdc++ does not support C++11, and libc++ is still
supported on our oldest version (10.8).

This copies configuration between cxx_stl_setup.mk and ninja.mk, but building
ckati and makeparallel with make will be removed soon.

Change-Id: I1fccee8da0f425e10ccc9d3247ed40664eb6ada0