Update patch set 3
Patch Set 3:

What are you planning on setting CCACHE_COMPILERCHECK to?

It sounds like the embedded kernel build should be resetting the value of CCACHE_BASEDIR, so that absolute paths (which are rarely used for android builds) are rewritten properly.

Or CCACHE_DIRECT should be turned off, so that header location changes are actually noticed -- this reduces the time savings, but fixes correctness problems such as this one.

Patch-set: 3