Revert "Revert "releasetools: Support OTAs that have OEM properties changes.""
This CL fixes the bug in [1] and reenables it.

We need to handle a special case that an OTA goes from a source build
without OEM properties to a target build with those properties (or vice
versa). Add support in OTA scripts to deal the case properly, by a)
using two oem_props variables to handle source and target builds
respectively; b) adjusting the fingerprint/thumbprint assertions to
allow a mix of both.

[1] commit c086370440bb32501da427718a7313f12ec2c841

Change-Id: I98118d77d5a0ff694fa1ee33602b5ee5e048599b