Update patch set 1
Patch Set 1:

Yeah, but this also removes the ability to install just one module (without knowing the exact installed filename).  (And yes, we already may install both the host and target modules -- I think that's likely a bug too)

I guess I was most worried that this would actually change builds too if people used a module name in LOCAL_ADDITIONAL_DEPENDENCIES, but that's already wrong and causes other problems.

In an ideal world, if the module was already being installed onto the target, I'd like specifying it to only build/install the variants that would normally be installed. And have the other variants be explicit options. That fixes the problem of accidentally installing too many things to the system image that wouldn't normally be there. But I don't think that's feasible before Soong.

Patch-set: 1