Update patch set 1
Patch Set 1:

For the detail, I've just sent an email to the thread about incremental build, but a short description: I was wondering if we can eliminate all dependency issues which potentially cause incremental build failures. I hacked a script which detects dependency issues, and this change is a demo of the tool.

Patch-set: 1
Label: Verified=0